refactor(settings): split settings.rs into modular submodules
- Delete `src-tauri/src/settings.rs` and create a modular directory structure under `src-tauri/src/settings/`.
- Move setting configurations, structs, and enums to `settings/model.rs`.
- Extract partial patch deserialization and deep-merging logic to `settings/patch.rs`.
- Relocate file-based persistence, atomic write helpers, and synchronization locks to `settings/provider.rs`.
- Move validation bounds tables, custom validators, and violation parsing logic to `settings/validation.rs`.
- Update `settings/mod.rs` to coordinate submodules and expose `AppSettings` public methods.
- Regenerate TypeScript bindings in `src/bindings/bindings.ts` to reflect updated crate path links in comments.
